imprv: move judgePiece method to PieceStore from index.js
